Looks like our IRQ's and interrupts are working its time to enter the scary territory of paging and task switching. (AAAARHRHRGHGH)
git-svn-id: http://gommo.homelinux.net/svn/repos/os/trunk@21 d997549a-3ad4-0310-b91a-fb2166dfab67

+++ lib/string/strlen.c
@@ -12,11 +12,13 @@
 * 
 * Note: MUCH OF THIS CODE IS BASED ON GPL CODE AVAILABLE FROM THE LIBC 
 *       LIBRARY AND THE LINUX KERNEL
+*       The actual functions needed by the C Library were referenced from
+*       http://www.acm.uiuc.edu/webmonkeys/book/c_guide/
 **************************************************************************/
 #include <string.h>
 
-unsigned int strlen( const char *str )
-{
+size_t strlen( const char *str )
+{   
 /*    const char *ret = str;
 
     while ( *str++ );
